Clean up root of the crate
